Version 1 - basic dockerized pipeline/requirements.txt
asyncio-nats-client==0.8.2
